# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal Agent System",
    "description": "企业级多Agent知识库助手系统",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tools/categories/list": {
      "get": {
        "tags": [
          "工具管理"
        ],
        "summary": "获取工具分类列表",
        "description": "获取所有工具的分类列表",
        "operationId": "get_tool_categories_api_v1_tools_categories_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tools": {
      "get": {
        "tags": [
          "工具管理"
        ],
        "summary": "获取所有工具列表",
        "description": "获取系统中所有可用工具的列表",
        "operationId": "get_tools_list_api_v1_tools_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/{tool_name}": {
      "get": {
        "tags": [
          "工具管理"
        ],
        "summary": "获取工具详情",
        "description": "获取指定工具的详细信息",
        "operationId": "get_tool_detail_api_v1_tools__tool_name__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tool_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/{tool_name}/execute": {
      "post": {
        "tags": [
          "工具管理"
        ],
        "summary": "执行工具",
        "description": "执行指定的工具",
        "operationId": "execute_tool_api_v1_tools__tool_name__execute_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tool_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolExecuteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolExecuteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ToolDetailResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "是否成功",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/ToolInfo",
            "description": "工具详情"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ToolDetailResponse",
        "description": "工具详情响应",
        "example": {
          "data": {
            "category": "language",
            "description": "多语言翻译工具",
            "name": "translation",
            "parameters": [],
            "timeout": 30
          },
          "success": true
        }
      },
      "ToolExecuteRequest": {
        "properties": {
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters",
            "description": "工具参数"
          }
        },
        "type": "object",
        "required": [
          "parameters"
        ],
        "title": "ToolExecuteRequest",
        "description": "工具执行请求",
        "example": {
          "parameters": {
            "source_lang": "en",
            "target_lang": "zh",
            "text": "Hello, world!"
          }
        }
      },
      "ToolExecuteResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "是否成功"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "执行结果"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "错误信息"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ToolExecuteResponse",
        "description": "工具执行响应",
        "example": {
          "data": {
            "translated_text": "你好，世界！"
          },
          "success": true
        }
      },
      "ToolInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "工具名称"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "工具描述"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "description": "工具分类"
          },
          "parameters": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Parameters",
            "description": "参数列表"
          },
          "timeout": {
            "type": "integer",
            "title": "Timeout",
            "description": "超时时间（秒）"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "category",
          "parameters",
          "timeout"
        ],
        "title": "ToolInfo",
        "description": "工具信息",
        "example": {
          "category": "language",
          "description": "多语言翻译工具",
          "name": "translation",
          "parameters": [
            {
              "description": "要翻译的文本",
              "name": "text",
              "required": true,
              "type": "string"
            }
          ],
          "timeout": 30
        }
      },
      "ToolListResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "是否成功",
            "default": true
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ToolInfo"
            },
            "type": "array",
            "title": "Data",
            "description": "工具列表"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "工具总数"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total"
        ],
        "title": "ToolListResponse",
        "description": "工具列表响应",
        "example": {
          "data": [],
          "success": true,
          "total": 9
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ErrorDetail": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Type"
          }
        },
        "required": [
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "example": {
          "code": 400,
          "details": [
            {
              "field": "email",
              "message": "Invalid email",
              "type": "value_error"
            }
          ],
          "error": "ValidationError",
          "error_code": "SYSTEM_VALIDATION_ERROR",
          "message": "Validation error",
          "timestamp": "2024-01-01T00:00:00+00:00"
        },
        "properties": {
          "code": {
            "title": "Code",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Details"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "error",
          "error_code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}